loader/requirements.txt
pandas==2.2.2
openpyxl==3.1.5
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
